Backhand finished

+++ src/main/java/ro/utcn/pt/lab01/Monom.java
@@ -3,7 +3,7 @@ package ro.utcn.pt.lab01;
 
 public class Monom {
 
-	int coef;
+	float coef;
 	int degree;
 	Monom next;
